Prettier: No trailing comma
This became the default in 2.0, and while there are good reasons for it,
I don't visually like it. It feels like extra clutter in the code, and
if I am striving to patch Prettier's format, another character to type.
Plus few of my projects have this on already, which requires a total
prettier fix.